\name{00.numDeriv.Intro}
\alias{00.numDeriv.Intro}
\docType{package}

\title{Accurate Numerical Derivatives}

\description{Calculate (accurate) numerical approximations to derivatives.}

\details{ See \code{\link{numDeriv-package}} ( in the help system use
   package?numDeriv or ?"numDeriv-package") for an overview.
}
\keyword{package}

